+++ pyout/elements.py
@@ -1,12 +1,7 @@
 """Style elements and schema validation.
 """
 
-from __future__ import unicode_literals
-
-try:
-    from collections.abc import Mapping
-except ImportError:  # Python <= 3.3
-    from collections import Mapping
+from collections.abc import Mapping
 
 schema = {
     "definitions": {